Making progress on new kernel system
